The Monster Begins
I'm building a monster and this is the first step.

Bots that can use AI APIs to think, make decisions, extend themselves, etc.

This system took a TON of time and effort to develop. I'm going to be working out how to make it extendable and grow on its own over time.